使ってないファイル削除。 コード修正。
