+++ src/components/Molecules/Swiper.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Swiper, SwiperSlide } from 'swiper/react'
 import SwiperCore, { Navigation, Pagination, EffectFade, Autoplay } from 'swiper'
 SwiperCore.use([Navigation])
@@ -6,8 +5,6 @@ import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import 'swiper/css/scrollbar';
-import style from "../../src/styles/swiper.module.css"
-
 
 SwiperCore.use([Pagination, Autoplay, EffectFade]);
