+++ src/components/ScrollToTop.tsx
@@ -2,13 +2,13 @@ import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 
 function ScrollTop() {
-　　const { pathname } = useLocation();
+  const { pathname } = useLocation();
 
-　　　　useEffect(() => {
-　　　　　　window.scrollTo(0, 0);
-　　　　}, [pathname]);
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, [pathname]);
 
-　　　　return null;
+  return null;
 }
 
 export default ScrollTop;
