+++ src/components/layout/Header.tsx
@@ -2,12 +2,11 @@ import { Link } from "react-router-dom";
 import headerStyle from "../../styles/layout/_Header.module.scss";
 import PrimaryButton from "../Atoms/button/PrimaryButton";
 import { useAuthState } from "react-firebase-hooks/auth";
-import { auth, provider } from "../../Firebase";
+import { auth } from "../../Firebase";
 import { GiFrogPrince } from "react-icons/gi";
 import Hamburger from "hamburger-react";
 import { useState } from "react";
 import { useEffect } from "react";
-import { Helmet } from "react-helmet";
 import { useNavigate } from "react-router-dom";
 
 const Header = () => {
@@ -33,7 +32,7 @@ const Header = () => {
       }
     });
 
-    // ex.) hJ2JnzBn=1ori4kgk 
+    // ex.) hJ2JnzBn=1ori4kgk
     // gestID（hJ2JnzBn）が入っていなければ、ランダムな文字列をcookieに追加
     if (cookieList.length === 0) {
       let randomId = Math.random().toString(32).substring(2);
@@ -53,7 +52,7 @@ const Header = () => {
 
   return (
     <>
-    {/* 上部移動用のID追加（react-scroll） */}
+      {/* 上部移動用のID追加（react-scroll） */}
       <div className={headerStyle.headerFlex} id="top">
         <Link to={"/"}>
           <h1>
@@ -78,22 +77,17 @@ const Header = () => {
                   </span>
                 </Link>
               </li>
-              {/* <li>
-              <Link to={"/"}>ログイン</Link>
-            </li> */}
               <Certification />
               <Link to={"/rooms/Gestroom"}>
                 <PrimaryButton>ご予約</PrimaryButton>
               </Link>
             </ul>
-            {/* <div className={headerStyle.ham}> */}
             <Hamburger
               toggled={isOpen}
               toggle={setOpen}
               size={30}
               onToggle={flagChange}
             />
-            {/* </div> */}
           </div>
         ) : (
           <>
